+++ tests/test_static.py
@@ -27,6 +27,38 @@ from pyspartn.spartnhelpers import (
     timetag2date,
     valid_crc,
 )
+from pyspartn.spartntypes_core import SPARTN_DATA_FIELDS, FL
+from pyspartn.spartntables import (
+    SF015_ENUM,
+    SF022_ENUM,
+    SF024_ENUM,
+    SF042_ENUM,
+    SF044_ENUM,
+    SF051_ENUM,
+    SF055_ENUM,
+    SF056_ENUM,
+    SF063_ENUM,
+    SF070_ENUM,
+    SF077_ENUM,
+    SF078_ENUM,
+    SF081_ENUM,
+    SF085_ENUM,
+    SF087_ENUM,
+    SF090_ENUM,
+    SF091_ENUM,
+    SF093_ENUM,
+    SF094_ENUM,
+    SF095_ENUM,
+    SF096_ENUM,
+    SF097_ENUM,
+    SF098_ENUM,
+    SF102_ENUM,
+    SF103_ENUM,
+    SF104_ENUM,
+    SF105_ENUM,
+    SF106_ENUM,
+    SF107_ENUM,
+)
 
 
 class StaticTest(unittest.TestCase):
@@ -44,6 +76,16 @@ class StaticTest(unittest.TestCase):
         bm = b"\x01\x08\x03\xf0\xff"
         for i, (ps, ln) in enumerate(bits):
             res = bitsval(bm, ps, ln)
+            self.assertEqual(res, EXPECTED_RESULT[i])
+
+    def testbitsval2(self):
+        bits = [(7, 1), (8, 8), (22, 2), (24, 4)]
+        EXPECTED_RESULT = [-0.5, 3.0, 0.5, 6.5]
+
+        bm = b"\x01\x08\x03\xf0\xff"
+        for i, (ps, ln) in enumerate(bits):
+            res = bitsval(bm, ps, ln, "FL", 0.5, -1.0)
+            # print(res)
             self.assertEqual(res, EXPECTED_RESULT[i])
 
     def testbitsvalerr(self):
@@ -164,6 +206,8 @@ class StaticTest(unittest.TestCase):
         self.assertEqual(res, "Ionosphere equation type")
         res = datadesc("SF043_01")
         self.assertEqual(res, "Area average vertical hydrostatic delay")
+        res = datadesc("SF049a")
+        self.assertEqual(res, "Large troposphere coefficient T01")
 
     def testenc2float(self):  # test enc2float
         res = enc2float(1332, 0.1, -90)
@@ -178,6 +222,50 @@ class StaticTest(unittest.TestCase):
     def testdate2timetag(self):  # test date2timetag
         res = date2timetag(datetime(2023, 6, 27, 21, 3, 0))
         self.assertEqual(res, 425595780)
+
+    def testdatafields(self):  # check float datafields are correctly configured
+        for _, value in SPARTN_DATA_FIELDS.items():
+            if value[1] == FL:
+                self.assertTrue(
+                    isinstance(value[3], (int, float))
+                    and isinstance(value[2], (int, float))
+                )
+
+    def testtables(self):  # sanity check on lookup tables
+        i = 0
+        for tbl in (
+            SF015_ENUM,
+            SF022_ENUM,
+            SF024_ENUM,
+            SF042_ENUM,
+            SF044_ENUM,
+            SF051_ENUM,
+            SF055_ENUM,
+            SF056_ENUM,
+            SF063_ENUM,
+            SF070_ENUM,
+            SF077_ENUM,
+            SF078_ENUM,
+            SF081_ENUM,
+            SF085_ENUM,
+            SF087_ENUM,
+            SF090_ENUM,
+            SF091_ENUM,
+            SF093_ENUM,
+            SF094_ENUM,
+            SF095_ENUM,
+            SF096_ENUM,
+            SF097_ENUM,
+            SF098_ENUM,
+            SF102_ENUM,
+            SF103_ENUM,
+            SF104_ENUM,
+            SF105_ENUM,
+            SF106_ENUM,
+            SF107_ENUM,
+        ):
+            i += len(tbl)
+        self.assertEqual(i, 141)
 
 
 if __name__ == "__main__":